Day 02
Power of Four
Given a non-negative integer input, return true if input is a power of 4. Return false otherwise.
  
Example #1:
Input: 16
Output: true 
Explanation: The number 16 is a power of four. (4^2)

Example #2:
Input: 8
Output: false 
Explanation: The number 8 is not a power of four. The exponent would be 3/2 which is not whole.

Constraints:  Your solution should run in O(1) time.