Day 22
Add Binary Strings
Given two strings s1 and s2 consisting of only ones and zeros, return a string representing the sum of s1 and s2 when they are considered as binary numbers.

Example #1:
Input: s1 = "101", s2 = "1"
Output: "110"
Explanation: In base 10, “101” represents 5, and “1” represents 1. Their sum is 6, which is “110” in binary.