Day 49
Given the root node to a binary tree, determine the longest path that exists between any two nodes in the tree.

Example 1:
Input:
    1

Output: 0
Explanation: Tree has a single node with no edges connecting it to another node. No longest node to node path can exist with only one node.
Example 2:
Input:
    1
   / 
  2  

Output: 1
Explanation: Paths [2,1] or [1,2] (the order of the nodes in sequence does not matter).   
Example 3:
Input:
    1
   / \
  2   3
     /  \
    4    5
   /    / \
  6    7   8 
   \
    9

Output: 5
Explanation:
Longest Paths:
- [9,6,4,3,1,2]
- [9,6,4,3,5,7]
- [9,6,4,3,5,8]

Any one of these would be a valid longest path in the binary tree.  The length of all these paths are 5.  Notice that the root node is not guaranteed to be on the longest path in a binary tree.