Day 60
Given an integer n, compute the square root of n, rounded down to the nearest integer.

Example #1:
Input: 25
Output: 5
Explanation: The square root of 25 is 5.
Example #2:
Input: 8
Output: 2
Explanation: Although the square root of 8 is 2.82, we round down to the nearest integer, so the answer is 2.
Constraints: Your solution should run in O(log(n)) time. You may not use the built-in square root function of your respective language.